Fully-deprecate closure-scoped views
Soft-deprecated in 1.5.0.
