+++ Sources/ComposableArchitecture/SwiftUI/Alert.swift
@@ -14,46 +14,8 @@ extension View {
     self._alert(store: store, state: { $0 }, action: { $0 })
   }
 
-  /// Displays an alert when the store's state becomes non-`nil`, and dismisses it when it becomes
-  /// `nil`.
-  ///
-  /// - Parameters:
-  ///   - store: A store that is focused on ``PresentationState`` and ``PresentationAction`` for an
-  ///     alert.
-  ///   - toDestinationState: A transformation to extract alert state from the presentation state.
-  ///   - fromDestinationAction: A transformation to embed alert actions into the presentation
-  ///     action.
-  @available(
-    iOS, deprecated: 9999,
-    message:
-      "Further scope the store into the 'state' and 'action' cases, instead. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
-  @available(
-    macOS, deprecated: 9999,
-    message:
-      "Further scope the store into the 'state' and 'action' cases, instead. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
-  @available(
-    tvOS, deprecated: 9999,
-    message:
-      "Further scope the store into the 'state' and 'action' cases, instead. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
-  @available(
-    watchOS, deprecated: 9999,
-    message:
-      "Further scope the store into the 'state' and 'action' cases, instead. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
   @preconcurrency @MainActor
-  public func alert<State, Action, ButtonAction>(
-    store: Store<PresentationState<State>, PresentationAction<Action>>,
-    state toDestinationState: @escaping (_ state: State) -> AlertState<ButtonAction>?,
-    action fromDestinationAction: @escaping (_ alertAction: ButtonAction) -> Action
-  ) -> some View {
-    self._alert(store: store, state: toDestinationState, action: fromDestinationAction)
-  }
-
-  @preconcurrency @MainActor
-  private func _alert<State, Action, ButtonAction>(
+  func _alert<State, Action, ButtonAction>(
     store: Store<PresentationState<State>, PresentationAction<Action>>,
     state toDestinationState: @escaping (_ state: State) -> AlertState<ButtonAction>?,
     action fromDestinationAction: @escaping (_ alertAction: ButtonAction) -> Action
