+++ Sources/ComposableArchitecture/SwiftUI/IfLetStore.swift
@@ -123,154 +123,6 @@ public struct IfLetStore<State, Action, Content: View>: View {
     }
   }
 
-  /// Initializes an ``IfLetStore`` view that computes content depending on if a store of
-  /// ``PresentationState`` and ``PresentationAction`` is `nil` or non-`nil`.
-  ///
-  /// - Parameters:
-  ///   - store: A store of optional state.
-  ///   - ifContent: A function that is given a store of non-optional state and returns a view that
-  ///     is visible only when the optional state is non-`nil`.
-  ///   - elseContent: A view that is only visible when the optional state is `nil`.
-  @available(
-    iOS, deprecated: 9999,
-    message:
-      "Scope the store into the destination's wrapped 'state' and presented 'action', instead: 'store.scope(state: \\.destination, action: \\.destination.presented)'. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
-  @available(
-    macOS, deprecated: 9999,
-    message:
-      "Scope the store into the destination's wrapped 'state' and presented 'action', instead: 'store.scope(state: \\.destination, action: \\.destination.presented)'. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
-  @available(
-    tvOS, deprecated: 9999,
-    message:
-      "Scope the store into the destination's wrapped 'state' and presented 'action', instead: 'store.scope(state: \\.destination, action: \\.destination.presented)'. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
-  @available(
-    watchOS, deprecated: 9999,
-    message:
-      "Scope the store into the destination's wrapped 'state' and presented 'action', instead: 'store.scope(state: \\.destination, action: \\.destination.presented)'. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
-  @preconcurrency @MainActor
-  public init<IfContent, ElseContent>(
-    _ store: Store<PresentationState<State>, PresentationAction<Action>>,
-    @ViewBuilder then ifContent: @escaping (_ store: Store<State, Action>) -> IfContent,
-    @ViewBuilder else elseContent: @escaping () -> ElseContent
-  ) where Content == _ConditionalContent<IfContent, ElseContent> {
-    self.init(
-      store.scope(state: \.wrappedValue, action: \.presented),
-      then: ifContent,
-      else: elseContent
-    )
-  }
-
-  /// Initializes an ``IfLetStore`` view that computes content depending on if a store of
-  /// ``PresentationState`` and ``PresentationAction`` is `nil` or non-`nil`.
-  ///
-  /// - Parameters:
-  ///   - store: A store of optional state.
-  ///   - ifContent: A function that is given a store of non-optional state and returns a view that
-  ///     is visible only when the optional state is non-`nil`.
-  @available(
-    iOS, deprecated: 9999,
-    message:
-      "Scope the store into the destination's wrapped 'state' and presented 'action', instead: 'store.scope(state: \\.destination, action: \\.destination.presented)'. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
-  @available(
-    macOS, deprecated: 9999,
-    message:
-      "Scope the store into the destination's wrapped 'state' and presented 'action', instead: 'store.scope(state: \\.destination, action: \\.destination.presented)'. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
-  @available(
-    tvOS, deprecated: 9999,
-    message:
-      "Scope the store into the destination's wrapped 'state' and presented 'action', instead: 'store.scope(state: \\.destination, action: \\.destination.presented)'. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
-  @available(
-    watchOS, deprecated: 9999,
-    message:
-      "Scope the store into the destination's wrapped 'state' and presented 'action', instead: 'store.scope(state: \\.destination, action: \\.destination.presented)'. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
-  @preconcurrency @MainActor
-  public init<IfContent>(
-    _ store: Store<PresentationState<State>, PresentationAction<Action>>,
-    @ViewBuilder then ifContent: @escaping (_ store: Store<State, Action>) -> IfContent
-  ) where Content == IfContent? {
-    self.init(
-      store.scope(state: \.wrappedValue, action: \.presented),
-      then: ifContent
-    )
-  }
-
-  /// Initializes an ``IfLetStore`` view that computes content depending on if a store of
-  /// ``PresentationState`` and ``PresentationAction`` is `nil` or non-`nil` and state can further
-  /// be extracted from the destination state, _e.g._ it matches a particular case of an enum.
-  ///
-  /// - Parameters:
-  ///   - store: A store of optional state.
-  ///   - toState: A closure that attempts to extract state for the "if" branch from the destination
-  ///     state.
-  ///   - fromAction: A closure that embeds actions for the "if" branch in destination actions.
-  ///   - ifContent: A function that is given a store of non-optional state and returns a view that
-  ///     is visible only when the optional state is non-`nil` and state can be extracted from the
-  ///     destination state.
-  ///   - elseContent: A view that is only visible when state cannot be extracted from the
-  ///     destination.
-  @available(
-    *, deprecated,
-    message:
-      "Further scope the store into the 'state' and 'action' cases, instead. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
-  @preconcurrency @MainActor
-  public init<DestinationState, DestinationAction, IfContent, ElseContent>(
-    _ store: Store<PresentationState<DestinationState>, PresentationAction<DestinationAction>>,
-    state toState: @escaping (_ destinationState: DestinationState) -> State?,
-    action fromAction: @escaping (_ action: Action) -> DestinationAction,
-    @ViewBuilder then ifContent: @escaping (_ store: Store<State, Action>) -> IfContent,
-    @ViewBuilder else elseContent: @escaping () -> ElseContent
-  ) where Content == _ConditionalContent<IfContent, ElseContent> {
-    self.init(
-      store.scope(
-        state: { $0.wrappedValue.flatMap(toState) },
-        action: { .presented(fromAction($0)) }
-      ),
-      then: ifContent,
-      else: elseContent
-    )
-  }
-
-  /// Initializes an ``IfLetStore`` view that computes content depending on if a store of
-  /// ``PresentationState`` and ``PresentationAction`` is `nil` or non-`nil` and state can further
-  /// be extracted from the destination state, _e.g._ it matches a particular case of an enum.
-  ///
-  /// - Parameters:
-  ///   - store: A store of optional state.
-  ///   - toState: A closure that attempts to extract state for the "if" branch from the destination
-  ///     state.
-  ///   - fromAction: A closure that embeds actions for the "if" branch in destination actions.
-  ///   - ifContent: A function that is given a store of non-optional state and returns a view that
-  ///     is visible only when the optional state is non-`nil` and state can be extracted from the
-  ///     destination state.
-  @available(
-    *, deprecated,
-    message:
-      "Further scope the store into the 'state' and 'action' cases, instead. For more information, see the following article: https://swiftpackageindex.com/pointfreeco/swift-composable-architecture/main/documentation/composablearchitecture/migratingto1.5#Enum-driven-navigation-APIs"
-  )
-  @preconcurrency @MainActor
-  public init<DestinationState, DestinationAction, IfContent>(
-    _ store: Store<PresentationState<DestinationState>, PresentationAction<DestinationAction>>,
-    state toState: @escaping (_ destinationState: DestinationState) -> State?,
-    action fromAction: @escaping (_ action: Action) -> DestinationAction,
-    @ViewBuilder then ifContent: @escaping (_ store: Store<State, Action>) -> IfContent
-  ) where Content == IfContent? {
-    self.init(
-      store.scope(
-        state: { $0.wrappedValue.flatMap(toState) },
-        action: { .presented(fromAction($0)) }
-      ),
-      then: ifContent
-    )
-  }
 
   public var body: some View {
     WithViewStore(
